# api/swagger/soccer-manager.swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    properties:
      data: {}
    type: object
...
  LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  PlayerPositionCode:
    enum:
    - GLK
    - DEF
    - MID
    - ATK
    type: string
    x-enum-varnames:
    - PlayerPositionCodeGLK
    - PlayerPositionCodeDEF
    - PlayerPositionCodeMID
    - PlayerPositionCodeATK
...
  PlayerResponse:
    properties:
      age:
        type: integer
      country_code:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      position_code:
        $ref: '#/definitions/PlayerPositionCode'
      price:
        type: integer
      team_id:
        type: integer
    type: object
  RefreshResponse:
    properties:
      access_token:
        type: string
    type: object
  RegisterRequest:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - password
    - role
    - username
    type: object
  RegisterResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  UpdatePlayerDataRequest:
    properties:
      country_code:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    required:
    - country_code
    - first_name
    - last_name
    type: object
...
  echo.HTTPError:
    properties:
      message: {}
    type: object
host: localhost:80
info:
  contact: {}
  description: Soccer manager api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Soccer Manager
  version: 1.0.0-alpha0
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user with username and password and returns JWT
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Login user
      tags:
      - auth
...
  /v1/auth/refresh:
    post:
      description: Get a new access token using refresh token from cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/RefreshResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Refresh access token
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns JWT tokens
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Register new user
      tags:
      - auth
...
  /v1/players:
    get:
      description: Returns a paginated list of all players
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/PlayerResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get all players
      tags:
      - players
  /v1/players/{player_id}:
    get:
      description: Returns a single player's data by ID
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/PlayerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get player by ID
      tags:
      - players
    put:
      consumes:
      - application/json
      description: Updates a player's data by ID
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: integer
      - description: Player data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdatePlayerDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Update player data
      tags:
      - players
...
  /v1/teams/{team_id}/players:
    get:
      description: Returns a paginated list of players for the specified team
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/PlayerResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get players by team ID
      tags:
      - players
...
  /v1/users/{user_id}/players:
    get:
      description: Returns a paginated list of players for the specified user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/PlayerResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get players by user ID
      tags:
      - players
...
schemes:
- http
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"